Some int/float conversion cleanup

+++ src/app/windstille_main.cpp
@@ -172,7 +172,7 @@ void
 WindstilleMain::init_modules()
 {
   if (debug) std::cout << "Initialising Fonts" << std::endl;
-  SoundManager::current()->set_master_volume(config.get_int("master-volume")/100.0f);
+  SoundManager::current()->set_master_volume(static_cast<float>(config.get_int("master-volume"))/100.0f);
   SoundManager::current()->enable_sound(config.get_bool("sound"));
   SoundManager::current()->enable_music(config.get_bool("music"));
